updated nlp
we have a LLM to give a response if it hasn't understood your request,
only issue with it is that it takes quite a while to give a response but other than that it works really well now,

I have also changed it so the text is a different colour when it is the LLM vs the NLP so the user knows that it is the LLM that has given the response.

I have also cleaned up the NLP so that it works like this:

NLP_main - this is the main python file which uses functions from the other python files
NLP_booking - this contains all the NLP only used for booking trains
NLP_predict - this contains all the NLP only used for predictions
NLP_functions - this contains functions used in both prediction and booking

and lastly I also added the similarity function that only allows the user to select stations from the csv and not any place that they choose, I haven't tested if it works for spelling mistakes yet because it would likely fail as the NER would likely not understand the word to wouldn't recognise it as anything significant so it wouldn't even get far enough to be given to the similarity function.